Document what happens to extra arguments passed to createLogger
`createLogger` appears in a number of examples, but it is not specifically documented on its own.  In particular, it's possible to pass extra options to `createLogger` which then show up in all your logs, but that behavior was not documented. 

This commit addresses that, as well as tagging a code section as JavaScript so syntax highlighting can work.