Allow 'time' field to be set with user function
I love the bunyan library, but it has always been troubling to me that the time is recorded in the log files as UTC time, to the millisecond. I would much rather be able to provide a function (passed in using options.time) that formats the time how I would like. This would allow log file timestamps to be read much more easily by human users. We already have this functionality in the CLI tool, so why not here?

I did everything to stick to the current API which says to set the 'time' property, but found it too cumbersome, especially because if you pass in more than one parameter to the logging function, only the first parameter's object properties will be added as fields of the logging object (the rest will be stringified into the 'msg' field). For example if I do `log.error(new Error('The system crashed.'), {time: myTimeFunction()});`, the time will no longer be added as a field of the log record; if I switch the parameters, unfortunately, the Error will no longer be outputted with the stack trace since it is stringified instead of being serialized by Bunyan. Please consider this issue, and if you would rather implement it in another way that is also ok, but let's find something that'll work in these sorts of instances! Thanks in advance!

Example usage:

    ...
    var logger = bunyan.createLogger({name: 'mylogger', streams: myStreamsObject, time: function() {
        return (new Date()).toLocaleString();
    });
    ...
    logger.info('example output');    //behaves same as expected