Support serializing bigint values
The available JSON serialization utilities do not natively support `stringify`ing `BigInt` values,
but they do support passing replacer arguments for serialization of custom or unsupported types.

In order not to crash when serializing BigInt values, we can pass a replacer function that substitutes
these values to a string representation. The chosen format is `12345n`, which aligns with most other
string representations, as well as the shorthand syntax in later versions of JavaScript.